backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
psycopg[binary]
pgvector
pydantic-settings
alibabacloud-oss-v2
beautifulsoup4
openai
langchain
langchain-openai
langgraph
langgraph-checkpoint-postgres
pillow
pymupdf
pypdf
pytesseract
python-docx
pytest
httpx
volcengine-python-sdk[ark]
PyJWT
